fix: 3 mypy errors — type annotation, float cast, None guard
- gyrokinetic_transport.py: _at() helper missing type annotation
- gk_eigenvalue.py: _kbm_growth_rate returns float() explicitly
- integrated_scenario.py: guard _phase_omega None before gk_natural_frequencies

+++ src/scpn_control/core/gk_eigenvalue.py
@@ -164,7 +164,7 @@ def _kbm_growth_rate(
     """
     # eps = 1e-4: regulariser for near-zero magnetic shear
     ballooning_drive = alpha_MHD * beta_e / (s_hat**2 + 1e-4)
-    return gamma_es * np.sqrt(1.0 + ballooning_drive)
+    return float(gamma_es * np.sqrt(1.0 + ballooning_drive))
 
 
 def _mtm_growth_rate(
